refactor: migrate to `jest` (#13)
* build(npm): update npm scripts, deps

* build(webpack): remove webpack test config

* refactor: migrate test config to `jest`

* refactor: centralize tsconfig

* docs: add jest badge to README

* refactor(core): refactor `jest` migration
