--- packages/@ngx-i18n-router/core/tests/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../../../tsconfig.json",
-  "compilerOptions": {
-    "module": "commonjs",
-    "declaration": false,
-    "typeRoots": [
-      "../../../../node_modules/@types"
-    ]
-  }
-}
