--- packages/@ngx-i18n-router/core/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "extends": "../../../tsconfig.json",
-  "files": [
-    "./index.ts",
-    "./src/models/i18n-router-settings.ts",
-    "./src/i18n-router.loader.ts",
-    "./src/i18n-router.service.ts",
-    "./src/i18n-router.pipe.ts"
-  ],
-  "angularCompilerOptions": {
-    "strictMetadataEmit": true,
-    "skipTemplateCodegen": true
-  }
-}
