+++ tsconfig.json
@@ -1,26 +1,23 @@
 {
   "compilerOptions": {
+    "target": "es5",
+    "module": "commonjs",
     "moduleResolution": "node",
     "emitDecoratorMetadata": true,
     "experimentalDecorators": true,
-    "module": "es2015",
-    "target": "es5",
     "noImplicitAny": true,
     "suppressImplicitAnyIndexErrors": true,
     "importHelpers": true,
-    "declaration": true,
-    "removeComments": true,
-    "typeRoots": [
-      "./node_modules/@types"
-    ],
     "lib": [
       "es2017",
       "dom"
     ]
   },
+  "include": [
+    "scripts/*.ts",
+    "packages/**/*.ts"
+  ],
   "exclude": [
     "node_modules"
-  ],
-  "compileOnSave": false,
-  "buildOnSave": false
+  ]
 }
